rename user_relations.py to user_tables.py
