+++ datajoint/__init__.py
@@ -1,7 +1,7 @@
 """
-DataJoint for Python is a framework for building data piplines using MySQL databases 
+DataJoint for Python is a framework for building data piplines using MySQL databases
 to represent pipeline structure and bulk storage systems for large objects.
-DataJoint is built on the foundation of the relational data model and prescribes a 
+DataJoint is built on the foundation of the relational data model and prescribes a
 consistent method for organizing, populating, and querying data.
 
 The DataJoint data model is described in https://arxiv.org/abs/1807.11104
@@ -64,7 +64,7 @@ logger.setLevel(log_levels[config['loglevel']])
 # ------------- flatten import hierarchy -------------------------
 from .connection import conn, Connection
 from .base_relation import FreeRelation, BaseRelation
-from .user_relations import Manual, Lookup, Imported, Computed, Part
+from .user_tables import Manual, Lookup, Imported, Computed, Part
 from .relational_operand import Not, AndList, U
 from .heading import Heading
 from .schema import Schema as schema
@@ -81,7 +81,7 @@ def create_virtual_module(module_name, schema_name, create_schema=False, create_
     :param module_name: displayed module name
     :param schema_name: name of the database in mysql
     :param create_schema: if True, create the schema on the database server
-    :param create_tables: if True, module.schema can be used as the decorator for declaring new 
+    :param create_tables: if True, module.schema can be used as the decorator for declaring new
     :return: the python module containing classes from the schema object and the table classes
     """
     module = ModuleType(module_name)
